# UserManagementService/openapi.yaml
openapi: 3.0.0
info:
  title: User Management Service
  version: 1.0.0
  description: Gerencia cadastrado, autenticação e perfis de usuários
tags:
  - name: Auth
    description: Cadastro e autenticação de usuários
paths:
  /authenticate:
    post:
      summary: Authenticate a user
      operationId: authenticateUser
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        "404":
          description: Not Found
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /register:
    post:
      summary: Register a new user
      operationId: registerUser
      tags:
        - Auth      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  type: string
                password:
                  type: string
                confirmPassword:
                  type: string
                bio:
                  type: string
                role:
                  type: string
                  enum:
                    - CLIENT
                    - TRAINER
                    - INFLUENCER
      responses:
        "201":
          description: CREATED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
        username:
          type: string
        bio:
          type: string
        role:
          type: string
          enum:
            - CLIENT
            - TRAINER
            - INFLUENCER